Refactor: Extract a base struct of `meta` to contain the basic methods
