+++ internal/config/config.go
@@ -2,17 +2,27 @@ package config
 
 import "github.com/Azure/aztfy/internal/resmap"
 
-type Config struct {
-	LogPath             string
-	MockClient          bool
-	SubscriptionId      string
-	ResourceMapping     resmap.ResourceMapping
-	ResourceGroupName   string
-	OutputDir           string
-	ResourceNamePattern string
-	Overwrite           bool
-	Append              bool
-	BatchMode           bool
-	BackendType         string
-	BackendConfig       []string
+type Config interface {
+	isConfig()
 }
+
+type CommonConfig struct {
+	SubscriptionId string
+	OutputDir      string
+	Overwrite      bool
+	Append         bool
+	BatchMode      bool
+	BackendType    string
+	BackendConfig  []string
+}
+
+type RgConfig struct {
+	CommonConfig
+
+	ResourceGroupName   string
+	ResourceMapping     resmap.ResourceMapping
+	ResourceNamePattern string
+	MockClient          bool
+}
+
+func (RgConfig) isConfig() {}
