+++ internal/test/append_test.go
@@ -57,10 +57,12 @@ resource "azurerm_resource_group" "test3" {
 
 	// Import the first resource group
 	aztfyDir := t.TempDir()
-	cfg := config.Config{
-		SubscriptionId:      os.Getenv("ARM_SUBSCRIPTION_ID"),
-		OutputDir:           aztfyDir,
-		BackendType:         "local",
+	cfg := config.RgConfig{
+		CommonConfig: config.CommonConfig{
+			SubscriptionId: os.Getenv("ARM_SUBSCRIPTION_ID"),
+			OutputDir:      aztfyDir,
+			BackendType:    "local",
+		},
 		ResourceNamePattern: "t1",
 	}
 	cfg.ResourceGroupName = d.RandomRgName() + "1"
