+++ internal/ui/importlist/importlist.go
@@ -20,13 +20,13 @@ import (
 )
 
 type Model struct {
-	c        meta.Meta
+	c        meta.RgMeta
 	listkeys listKeyMap
 
 	list list.Model
 }
 
-func NewModel(c meta.Meta, l meta.ImportList, idx int) Model {
+func NewModel(c meta.RgMeta, l meta.ImportList, idx int) Model {
 	// Build candidate words for the textinput
 	candidates := make([]string, 0, len(azurerm.ProviderSchemaInfo.ResourceSchemas))
 	for rt := range azurerm.ProviderSchemaInfo.ResourceSchemas {
